some fixes

+++ src/asmc.c
@@ -10,8 +10,8 @@ void outb(uint16 port, uint8 data) {
     asm volatile ("outb %0, %1" :: "a"(data), "Nd"(port));
 }
 
-uint8 inw(uint16 port) {
-    uint8 data;
+uint16 inw(uint16 port) {
+    uint16 data;
     asm volatile ("inw %1, %0" : "=a"(data) : "Nd"(port));
     return data;
 }
